Updated to 11.2 Postgres release

+++ src/postgres/src_backend_nodes_value.c
@@ -12,7 +12,7 @@
  *	  implementation of Value nodes
  *
  *
- * Copyright (c) 2003-2017, PostgreSQL Global Development Group
+ * Copyright (c) 2003-2018, PostgreSQL Global Development Group
  *
  *
  * IDENTIFICATION
@@ -28,7 +28,7 @@
  *	makeInteger
  */
 Value *
-makeInteger(long i)
+makeInteger(int i)
 {
 	Value	   *v = makeNode(Value);
